123123
1323123

+++ final-api/src/main/java/eu/it/academy/api/services/IUserService.java
@@ -7,5 +7,5 @@ import eu.it.academy.entities.User;
 @Service
 public interface IUserService {
 
-    User findUser(int id);
+    User findUser(Long id);
 }
